include computed fields in alle_onverwacht_resultaten
